requirements.txt
pyserial
python-dotenv
openai